+++ src/app/app-routing.module.ts
@@ -11,7 +11,7 @@ import { TorneoEsperaComponent } from './torneo-espera/torneo-espera.component';
 import { TorneoComponent } from './torneo/torneo.component';
 import { TosComponent } from './tos/tos.component';
 import { PartidaPrivadaComponent } from './partida-privada/partida-privada.component';
-import { ReglasPartidaComponent } from './partida-privada/partida-privada.component';
+
 
 const routes: Routes = [
   { path: "", component: AppComponent },
@@ -25,7 +25,6 @@ const routes: Routes = [
   { path: "torneoEspera/:id", component: TorneoEsperaComponent },
   { path: "leaderboard", component: LeaderboardComponent },
   { path: "partidaPrivada", component:PartidaPrivadaComponent},
-  { path: "reglasPartida", component:ReglasPartidaComponent},
 ];
 
 @NgModule({
