+++ src/app/menu-inicial/menu-inicial.component.html
@@ -21,7 +21,7 @@
   <div class="boton1">
     <a routerLink="/game">Partida rápida</a>
     <button mat-button (click)="pedirCodigo()">Unirse P.Privada</button>
-    <a href="/reglasPartida">Crear P.Privada</a>
+    <button mat-button (click)="crearPPrivada()">Crear P.Privada</button>
     <a href="/torneo"> Buscar Torneo</a>
   </div>
 </div>
