--- src/app/partida-privada/reglasPartida.html
@@ -1,71 +1,0 @@
-<div class="popup">
-    <h2 mat-dialog-title>Configuracion de la partida</h2>
-    <div class="center">
-        <form class="custom-form">
-            <input
-            type="number" max="6" min="2"
-            class="form-control"
-            (change)="changeNplayers($event)"
-            [value]="component.nJugadores"
-            />
-            <label> Nº de jugadores</label>
-        </form>
-    </div>
-    <form class="custom-form">
-        <h3>Reglas:</h3>
-        <input
-        type="checkbox"
-        [checked]="component.reglas[0]"
-        (change)="component.reglas[0] = !component.reglas[0]"
-        />
-        <label (click)="component.reglas[0] = !component.reglas[0]"> 0 Switch</label>
-        <br>
-        <input
-          type="checkbox"
-          [checked]="component.reglas[1]"
-          (change)="component.reglas[1] = !component.reglas[1]"
-        />
-        <label (click)="component.reglas[1] = !component.reglas[1]"> Crazy 7</label>
-        <br>
-        <input
-          type="checkbox"
-          [checked]="component.reglas[2]"
-          (change)="component.reglas[2] = !component.reglas[2]"
-        />
-        <label (click)="component.reglas[2] = !component.reglas[2]"> Progressive draw</label>
-        <br>
-        <input
-          type="checkbox"
-          [checked]="component.reglas[3]"
-          (change)="component.reglas[3] = !component.reglas[3]"
-        />
-        <label (click)="component.reglas[3] = !component.reglas[3]"> Chaos draw</label>
-        <br>
-        <input
-          type="checkbox"
-          [checked]="component.reglas[4]"
-          (change)="component.reglas[4] = !component.reglas[4]"
-        />
-        <label (click)="component.reglas[4] = !component.reglas[4]"> Block draw</label>
-        <br>
-        <input
-          type="checkbox"
-          [checked]="component.reglas[5]"
-          (change)="component.reglas[5] = !component.reglas[5]"
-        />
-        <label (click)="component.reglas[5] = !component.reglas[5]"> Repeat draw</label>
-        <br>
-    </form>
-    <div class="right">
-        <h3>Tiempo de turno:</h3>
-        <form class="custom-form">
-            <select (change)="changeTturno($event)" [value]="component.tiempoTurno">
-                <option value="10"> 10 segundos </option>
-                <option value="15"> 15 segundos </option>
-                <option value="20"> 20 segundos </option>
-            </select>
-        </form>
-    </div>
-
-    <button mat-button (click)="this.dialogRef.close();">Cerrar</button>
-</div>
